+++ x/bank/exported/exported.go
@@ -21,5 +21,6 @@ type (
 	// NOTE: This is used solely for migration of x/params managed parameters.
 	Subspace interface {
 		GetParamSet(ctx sdk.Context, ps ParamSet)
+		Get(ctx sdk.Context, key []byte, ptr interface{})
 	}
 )